+++ libft/libft.h
@@ -91,7 +91,7 @@ int					ft_wordcounter(char const *s, char c);
 void				ft_swap(int *a, int *b);
 int					ft_numlen(long long n);
 long long			ft_atol(const char *str);
-unsigned long long	ft_puissance(int nbr, int p);
+long				ft_puissance(int nbr, int p);
 void				ft_reverse(char **str);
 long				ft_long_atoi(const char *str);
 #endif
