+++ srcs/vm_get_values.c
@@ -12,11 +12,11 @@
 
 #include "../includes/vm.h"
 
-int			get_reg_val(t_all *vm, t_process *p, int bytes_read)
+long		get_reg_val(t_all *vm, t_process *p, int bytes_read)
 {
 	int		reg;
 
-	reg = get_next_bytes(vm, p, 1, bytes_read);
+	reg = (int)get_next_bytes(vm, p, 1, bytes_read);
 	if (reg < 1 || reg > REG_NUMBER)
 	{
 		p->op_fail = 1;
@@ -25,26 +25,26 @@ int			get_reg_val(t_all *vm, t_process *p, int bytes_read)
 	return (p->r[reg - 1]);
 }
 
-int			get_ind(t_all *vm, t_process *p, int bytes_read, int restriction)
+long			get_ind(t_all *vm, t_process *p, int bytes_read, int restriction)
 {
 	if (restriction == 1)
 		return (get_next_bytes(vm, p, 2, bytes_read) % IDX_MOD);
 	return (get_next_bytes(vm, p, 2, bytes_read));
 }
 
-int			get_val_at_ind(t_all *vm, t_process *p, int bytes_read,
+long			get_val_at_ind(t_all *vm, t_process *p, int bytes_read,
 			int restriction)
 {
-	int		address;
+	long	address;
 
 	address = get_ind(vm, p, bytes_read, restriction);
 	return (get_next_bytes(vm, p, 4, address - 1));
 }
 
-int			get_unspecified_val(t_all *vm, t_process *p, int *bytes_read,
+long		get_unspecified_val(t_all *vm, t_process *p, long *bytes_read,
 			int param)
 {
-	int		val;
+	long	val;
 
 	if (p->decode[param] == REG_CODE)
 	{
@@ -53,7 +53,9 @@ int			get_unspecified_val(t_all *vm, t_process *p, int *bytes_read,
 	}
 	if (p->decode[param] == IND_CODE)
 	{
+	
 		val = get_val_at_ind(vm, p, bytes_read[0], 1);
+		
 		bytes_read[0] += 2;
 	}
 	if (p->decode[param] == DIR_CODE)
@@ -64,10 +66,10 @@ int			get_unspecified_val(t_all *vm, t_process *p, int *bytes_read,
 	return (val);
 }
 
-int			get_unspecified_val_2(t_all *vm, t_process *p, int *bytes_read,
+long		get_unspecified_val_2(t_all *vm, t_process *p, long *bytes_read,
 			int param)
 {
-	int		val;
+	long	val;
 
 	if (p->decode[param] == REG_CODE)
 	{
